Move filters into separate package
