+++ cuppa/src/main/java/org/forgerock/cuppa/internal/filters/OnlyTestBlockFilter.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2015-2016 ForgeRock AS.
+ * Copyright 2016 ForgeRock AS.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.forgerock.cuppa.internal;
+package org.forgerock.cuppa.internal.filters;
 
 import static org.forgerock.cuppa.model.Behaviour.ONLY;
